requirements.txt
discord.py==2.3.2
etherscan_python==2.1.0
python-dotenv==1.0.0
python-telegram-bot==20.7
Requests==2.31.0
web3==6.13.0
